tools/payment-tool/requirements.txt
# The web server

fastapi
uvicorn

# The PostgreSQL database driver

psycopg2-binary